CIS 651 Week 1 Demo
Single activity where user is asked to enter a number to find a radnomly generated intiger in 10 attempts.
User is given hint to go up or down depending on the number they entered. Progress bar indicates the number
of attempts left and the botton linear layout includes the numbers user entered as history where colors
indicate if the number was bigger or smaller than the target. Final score is given as 10 * [number of attempts left].
This also includes a seperate landscape orientation where is history is located at the right side of the screen vertically.